restaurant_voting_api/requirements.txt
Django==5.1.4
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.10
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.9.0
